Temporarily remove TT's dependence on JES_EVIOpp
DTranslationTable.cc depends on both JEventSource_EVIO and JEventSource_EVIOpp, specifically in order to report back a set of ROCIDs to parse. This extra coupling makes DTranslationTable much harder to work with, and isn't even necessary. However, it is difficult to perform the surgery to reverse the direction of dependence when JEventSource_EVIO and JEventSource_EVIOpp contain hundreds of compiler errors. Instead, we are going to comment out the dependence on JEventSource_EVIOpp, finish the port of JEventSource_EVIO as it is, and refactor later.