DEvent is now a namespace
DEvent used to be an anemic object for providing convenience methods around JEvent and JApplication, coupled with a collection of free functions which had the same functionality, minus a tiny bit of dubious caching. The anemic object is now gone, and the free functions have been moved into the DEvent namespace.