Remove CMake infrastructure
While GlueX may eventually want to migrate from SCons to CMake, this implementation is insufficient for that purpose. Since nobody is using CLion, it is not necessary to finish the port. Rather than add noise to the codebase, we remove it until we are ready to add CMake support properly.